+++ src/payload.config.ts
@@ -118,6 +118,7 @@ const resendAdapter: EmailAdapter = () => {
 };
 
 export default buildConfig({
+  serverURL: process.env.NEXT_PUBLIC_SERVER_URL || '',
   admin: { user: Users.slug },
   collections: [
     Users,